Move machine-controller to subdir
